Remove most config files from dist-git and take them from sources
The content of these files is more or less tied to the policy source
code. Therefore, moving these files to the source repo rather than
dist-git will make it easier to do changes that would formerly need
coordinated modification both in the sources and in dist-git (e.g.
adding or removing a module). It will also make it easier for other
distributions seeking to package a Fedora-like SELinux policy.

[skip changelog]

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>