Use RPM's build conditionals to control which policy types to build
Instead of plain macros, use `%bcond ...` and `%{with ...}`, which will
allow controlling which policy types to build using the --with/--without
command-line arguments when calling `rpmbuild` or `mock`.

See also:
https://rpm-software-management.github.io/rpm/manual/conditionalbuilds.html

Note that the BUILD_DOC macro is removed without replacement as it's
unused. (The builds of the -doc and -devel subpackages overlap too much
for the macro to be useful, anyway.)

Signed-off-by: Ondrej Mosnáček <omosnacek@gmail.com>