Use `Requires(meta): (rpm-plugin-selinux if rpm-libs)`
We support two ways to update the operating system:

- `/usr/bin/rpm` (and `dnf` etc.) where SELinux labels are
  computed and written client side
- ostree (and other image-based systems) where SELinux labels
  were computed server side.

In the ostree case, I'd like the ability to generate smaller
images that do not even have `rpm` installed.

This hard dependency from `selinux-policy` -> `rpm` is one of
the only main blockers.

RPM supports these "alternative" conditionals, it's easy to do.