split the tests into 2 groups
When all relevant tests are executed on 1 machine, it may take
a significant time to get the test results. Because the Testing
Farm is capable to run multiple test plans in parallel, we can
split the relevant tests and run them on 2 machines. As a result,
we will get the test results sooner.