Drop the publicfile module
Fedora does not ship any package with /usr/bin/ftpd nor /usr/bin/httpd
and no explicit transition from other modules exists, so the content of
this module is ineffective.

Furthermore, an entry for /usr/bin/httpd clashes with a rule for the
same path in the apache.fc module after the bin-sbin merge, so the rule
in publicfile.fc now overrides the one in apache.fc which actually
needs to be used.